Player: widen primary stats from uint8_t to uint16_t
m_maxSTR..m_maxCHA were declared uint8_t, capping at 255. Modern Live
characters routinely exceed that — e.g. a level-65 caster with raid
gear hits INT 330+. Truncated INT (74 instead of 330) made
calcMaxMana undershoot dramatically, and the protoencoder's
mana_max = max(curMana, calcMaxMana) fallback then snapped the
displayed bar to "cur / cur" instead of cur / actual_max.

Widening to uint16_t fixes the truncation. The save/restore path
serializes these via QDataStream by static type, so the on-disk
format width changes; bump the magic from "plr2" to "plr3" so any
existing lastPlayer.dat rejects cleanly via the existing magic
check rather than mis-deserializing every following field.

Note: this does not by itself produce a correct mana_max readout
(calcMaxMana is still missing item / AA / modern-era contributions),
but it removes one of the two compounding bugs and unblocks the
search for the rest.

+++ src/player.h
@@ -131,13 +131,13 @@ public:
    int getPlusHP() { return m_plusHP; }
    int getPlusMana() { return m_plusMana; }
 
-   uint8_t getMaxSTR() { return m_maxSTR; }
-   uint8_t getMaxSTA() { return m_maxSTA; }
-   uint8_t getMaxCHA() { return m_maxCHA; }
-   uint8_t getMaxDEX() { return m_maxDEX; }
-   uint8_t getMaxINT() { return m_maxINT; }
-   uint8_t getMaxAGI() { return m_maxAGI; }
-   uint8_t getMaxWIS() { return m_maxWIS; }
+   uint16_t getMaxSTR() { return m_maxSTR; }
+   uint16_t getMaxSTA() { return m_maxSTA; }
+   uint16_t getMaxCHA() { return m_maxCHA; }
+   uint16_t getMaxDEX() { return m_maxDEX; }
+   uint16_t getMaxINT() { return m_maxINT; }
+   uint16_t getMaxAGI() { return m_maxAGI; }
+   uint16_t getMaxWIS() { return m_maxWIS; }
    uint16_t getMaxMana() { return m_maxMana; }
    uint16_t getMana() { return m_mana; }
    uint32_t getSpellBookSlot(uint32_t slotid) { return m_spellBookSlots[slotid]; }
@@ -249,15 +249,18 @@ public:
   
   uint16_t m_plusMana;
   uint16_t m_plusHP;
-  
+
   uint16_t m_maxMana;
-  uint8_t m_maxSTR;
-  uint8_t m_maxSTA;
-  uint8_t m_maxCHA;
-  uint8_t m_maxDEX;
-  uint8_t m_maxINT;
-  uint8_t m_maxAGI;
-  uint8_t m_maxWIS;
+  // Widened from uint8_t (legacy showeq cap of 255) to fit modern Live's
+  // post-cap stats — players today routinely exceed 255 (e.g. INT 330)
+  // and the truncation was making calcMaxMana undershoot wildly.
+  uint16_t m_maxSTR;
+  uint16_t m_maxSTA;
+  uint16_t m_maxCHA;
+  uint16_t m_maxDEX;
+  uint16_t m_maxINT;
+  uint16_t m_maxAGI;
+  uint16_t m_maxWIS;
   
   uint16_t m_food;
   uint16_t m_water;
